+++ CHANGES.rst
@@ -1,7 +1,8 @@
 0.7.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Update tests to adapt to changes in Redshift and SQLAlchemy
+  (`Issue #140 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/140>`_)
 
 
 0.7.1 (2018-01-17)